Cleanup settings utils file
